Move handler functions to the _handlers module
